evaluation/VLMEvalKit/requirements.txt
accelerate
dotenv
einops
# for gemini api
google-genai
gradio
huggingface_hub
imageio
ipdb
json_repair
matplotlib
nltk
numpy
omegaconf
openai
opencv-python>=4.7.0.72
openpyxl
pandas
pillow
portalocker
protobuf
python-dotenv
qwen_vl_utils
requests
rich
sentencepiece
setuptools
sty
tabulate
tiktoken
timeout-decorator
timm
torch
torchvision
tqdm
# Keep aligned with repository runtime versions to avoid tokenizer parse
# incompatibilities when this toolkit is installed in the same environment.
transformers==4.47.1
tokenizers==0.21.4
typing_extensions
validators
xlsxwriter
